feat(validation): add ParseIntPipe for request parameter validation

+++ src/articles/articles.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateArticleDto } from './dto/create-article.dto';
 import { UpdateArticleDto } from './dto/update-article.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -35,15 +39,18 @@ export class ArticlesService {
       const tagEntities = await this.tagRepository.find({
         where: { name: In(tags) },
       });
+
       articleTagsEntities.push(...tagEntities);
 
       const existingTagNames = tagEntities.map((tag) => tag.name);
+
       const newTags = tags.filter((tag) => !existingTagNames.includes(tag));
 
       if (newTags.length > 0) {
         const newTagEntities = this.tagRepository.create(
           newTags.map((name) => ({ name })),
         );
+
         articleTagsEntities.push(...newTagEntities);
         await this.tagRepository.save(newTagEntities);
       }
@@ -96,6 +103,7 @@ export class ArticlesService {
     }
 
     const article = await query.getOne();
+
     if (!article) {
       throw new NotFoundException(`Статья с ID ${id} не найдена`);
     }
@@ -104,10 +112,16 @@ export class ArticlesService {
   }
 
   async update(
-    id: number,
+    articleId: number,
+    userId: number,
     updateArticleDto: UpdateArticleDto,
   ): Promise<Article> {
-    const article = await this.findOne(id, true);
+    const article = await this.findOne(articleId, true);
+
+    if (article.author.id !== userId) {
+      throw new ForbiddenException('1');
+    }
+
     Object.assign(article, updateArticleDto);
 
     if (updateArticleDto.tags) {
@@ -131,9 +145,14 @@ export class ArticlesService {
     return this.articleRepository.save(article);
   }
 
-  async remove(id: number): Promise<void> {
-    const article = await this.findOne(id, true);
-    await this.articleRepository.remove(article);
+  async remove(articleId: number, userId: number) {
+    const article = await this.findOne(articleId, true);
+
+    if (article.author.id !== userId) {
+      throw new ForbiddenException('1');
+    }
+
+    return this.articleRepository.remove(article);
   }
 
   queryBuilder() {
